+++ packages/core/components/login-screen/login-screen-class.js
@@ -9,7 +9,7 @@ class LoginScreen extends Modal {
     }, params);
 
     // Extends with open/close Modal methods;
-    super(app, extendedParams);
+    super(extendedParams, app);
 
     const loginScreen = this;