+++ packages/core/components/popup/popup-class.js
@@ -12,7 +12,7 @@ class Popup extends Modal {
     );
 
     // Extends with open/close Modal methods;
-    super(app, extendedParams);
+    super(extendedParams, app);
 
     const popup = this;