+++ packages/react/components/list-button.d.ts
@@ -1,4 +1,5 @@
 import * as React from 'react';
+import { Tooltip as TooltipNamespace } from 'framework7/components/tooltip/tooltip';
 
 declare namespace F7ListButton {
   interface Props {
@@ -15,6 +16,7 @@ declare namespace F7ListButton {
     link? : boolean | string
     href? : boolean | string
     target? : string
+    tooltip? : string
     color? : string
     colorTheme? : string
     textColor? : string
@@ -64,5 +66,6 @@ declare namespace F7ListButton {
 }
 declare class F7ListButton extends React.Component<F7ListButton.Props, {}> {
   onClick(event? : any) : unknown
+  f7Tooltip: TooltipNamespace.Tooltip
 }
 export default F7ListButton;