+++ packages/react/components/list-button.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import Utils from '../utils/utils';
 import Mixins from '../utils/mixins';
+import __reactComponentWatch from '../runtime-helpers/react-component-watch.js';
 import __reactComponentDispatchEvent from '../runtime-helpers/react-component-dispatch-event.js';
 import __reactComponentSlots from '../runtime-helpers/react-component-slots.js';
 import __reactComponentSetProps from '../runtime-helpers/react-component-set-props.js';
@@ -80,9 +81,21 @@ class F7ListButton extends React.Component {
     const linkEl = self.refs.linkEl;
     linkEl.removeEventListener('click', this.onClick);
     delete linkEl.f7RouteProps;
+
+    if (self.f7Tooltip && self.f7Tooltip.destroy) {
+      self.f7Tooltip.destroy();
+      self.f7Tooltip = null;
+      delete self.f7Tooltip;
+    }
   }
 
-  componentDidUpdate() {
+  componentDidUpdate(prevProps, prevState) {
+    __reactComponentWatch(this, 'props.tooltip', prevProps, prevState, newText => {
+      const self = this;
+      if (!newText || !self.f7Tooltip) return;
+      self.f7Tooltip.setText(newText);
+    });
+
     const self = this;
     const linkEl = self.refs.linkEl;
     const {
@@ -98,7 +111,8 @@ class F7ListButton extends React.Component {
     const self = this;
     const linkEl = self.refs.linkEl;
     const {
-      routeProps
+      routeProps,
+      tooltip
     } = self.props;
 
     if (routeProps) {
@@ -106,6 +120,14 @@ class F7ListButton extends React.Component {
     }
 
     linkEl.addEventListener('click', self.onClick);
+    self.$f7ready(f7 => {
+      if (tooltip) {
+        self.f7Tooltip = f7.tooltip.create({
+          targetEl: linkEl,
+          text: tooltip
+        });
+      }
+    });
   }
 
   get slots() {
@@ -136,7 +158,8 @@ __reactComponentSetProps(F7ListButton, Object.assign({
   tabLinkActive: Boolean,
   link: [Boolean, String],
   href: [Boolean, String],
-  target: String
+  target: String,
+  tooltip: String
 }, Mixins.colorProps, Mixins.linkRouterProps, Mixins.linkActionsProps));
 
 F7ListButton.displayName = 'f7-list-button';