+++ packages/vue/components/list-button.js
@@ -15,7 +15,8 @@ export default {
     tabLinkActive: Boolean,
     link: [Boolean, String],
     href: [Boolean, String],
-    target: String
+    target: String,
+    tooltip: String
   }, Mixins.colorProps, Mixins.linkRouterProps, Mixins.linkActionsProps),
 
   render() {
@@ -81,6 +82,13 @@ export default {
     }
 
   },
+  watch: {
+    'props.tooltip': function watchTooltip(newText) {
+      const self = this;
+      if (!newText || !self.f7Tooltip) return;
+      self.f7Tooltip.setText(newText);
+    }
+  },
 
   created() {
     Utils.bindMethods(this, ['onClick']);
@@ -90,7 +98,8 @@ export default {
     const self = this;
     const linkEl = self.$refs.linkEl;
     const {
-      routeProps
+      routeProps,
+      tooltip
     } = self.props;
 
     if (routeProps) {
@@ -98,6 +107,14 @@ export default {
     }
 
     linkEl.addEventListener('click', self.onClick);
+    self.$f7ready(f7 => {
+      if (tooltip) {
+        self.f7Tooltip = f7.tooltip.create({
+          targetEl: linkEl,
+          text: tooltip
+        });
+      }
+    });
   },
 
   updated() {
@@ -117,6 +134,12 @@ export default {
     const linkEl = self.$refs.linkEl;
     linkEl.removeEventListener('click', this.onClick);
     delete linkEl.f7RouteProps;
+
+    if (self.f7Tooltip && self.f7Tooltip.destroy) {
+      self.f7Tooltip.destroy();
+      self.f7Tooltip = null;
+      delete self.f7Tooltip;
+    }
   },
 
   methods: {